requirements.txt
# Core framework
fastapi==0.100.0
uvicorn==0.22.0

# Database
sqlalchemy==2.0.19
psycopg2-binary==2.9.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# Data validation
pydantic==2.1.1
pydantic-settings==2.0.3
email-validator==2.0.0

# Testing
pytest==7.4.0
httpx==0.24.1

# Rate limiting
slowapi==0.1.8